+++ core/src/main/scala/com/krux/hyperion/activity/package.scala
@@ -1,5 +1,7 @@
 package com.krux.hyperion
 
+import java.net.URI
+
 import com.krux.hyperion.common.S3Uri
 import com.krux.hyperion.common.S3Uri._
 
@@ -11,4 +13,5 @@ package object activity {
 
   final val EmrCommandRunner: String = "command-runner.jar"
 
+  final val EmrHadoopJarsDir: URI = URI.create("/var/lib/aws/emr/step-runner/hadoop-jars/")
 }