+++ CHANGELOG.md
@@ -6,6 +6,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - [#543](https://github.com/krux/hyperion/issues/543) - Replace joda time with java.time
 
+## 5.8.1 - 2019-12-06
+### Fixed
+- [#441](https://github.com/krux/hyperion/issues/441) - Fix SparkTaskActivity to correctly run with the command-runner.jar.
+
+## 5.8.0 - 2019-11-18
+### Changed
+- [#341](https://github.com/krux/hyperion/issues/341) - Generate activity dependencies without any transitive dependencies.
+
 ## 5.7.2 - 2019-11-08
 ### Changed
 - [#595](https://github.com/krux/hyperion/issues/595) - Not able to import compressionFormat enum in other repository which uses hyperion